ljy_requirements.txt
pysam
biopython
matplotlib
seaborn
pandas
biopython